Fixed GCC_WARN_64_TO_32_BIT_CONVERSION warnings

+++ SSZipArchive/minizip/ioapi.c
@@ -192,7 +192,7 @@ static long ZCALLBACK fseek64_file_func (voidpf  opaque, voidpf stream, ZPOS64_T
     }
     ret = 0;
 
-    if(fseeko64((FILE *)stream, offset, fseek_origin) != 0)
+    if(fseeko64((FILE *)stream, (long)offset, fseek_origin) != 0)
                         ret = -1;
 
     return ret;
